refactor: split large modules under file-size soft cap
Splits the four files that exceeded the AI-readiness 1500-line soft
limit into sibling sub-modules, following the pattern documented in
CLAUDE.md ("Adding a new top-level command"). All public names are
re-exported so existing imports keep working.

  cli.py (1947 -> 1496):
    + cli_appcompat.py — appcompat command + 2 helpers
    + cli_stack.py     — deps + stack-check commands
    + cli_suggest.py   — suggest-suppressions command

  diff_platform.py (1633 -> 1465):
    + diff_platform_templates.py — template inner-type detectors
    - dead-code: removed an unused duplicate of _diff_advanced_dwarf
      (the live one is in checker.py)

  dumper.py (1709 -> 1144):
    + dumper_castxml.py — _CastxmlParser class + vtable helpers

  compat/cli.py (1511 -> 1429):
    + compat/_errors.py — error classification helpers

Side-effect registration of the new CLI sub-modules is wired up at the
bottom of cli.py; the import-cycle allowlist in check_ai_readiness.py
and the disallow_untyped_decorators override in pyproject.toml are
updated for each new module.

Result: AI-readiness gate is now 0 errors, 0 warnings (down from 4 file-
size warnings). All 6259 fast unit tests pass, mypy reports no new
errors, ruff is clean.
