+++ abicheck/cli.py
@@ -15,17 +15,15 @@
 """CLI — abicheck dump | compare | compat (dump | check)."""
 from __future__ import annotations
 
-import json
 import logging
 import re
 import sys
-from collections.abc import Callable
 from pathlib import Path
 from typing import TYPE_CHECKING
 
 import click
 
-from .checker import DiffResult, LibraryMetadata, Verdict, compare
+from .checker import DiffResult, LibraryMetadata, compare
 from .compat.abicc_dump_import import import_abicc_perl_dump, looks_like_perl_dump
 from .compat.cli import compat_group
 from .dumper import dump
@@ -33,7 +31,6 @@ from .errors import AbicheckError
 from .serialization import load_snapshot, snapshot_to_json
 
 if TYPE_CHECKING:
-    from .appcompat import AppRequirements
     from .checker_types import DiffResult
     from .debug_resolver import DebugArtifact
     from .policy_file import PolicyFile
@@ -1440,457 +1437,6 @@ def compare_cmd(
     _exit_with_severity_or_verdict(result, sev_config, severity_explicitly_set)
 
 
-def _validate_appcompat_args(
-    weak_mode: bool,
-    old_lib: Path | None, new_lib: Path | None,
-    list_symbols: bool,
-    old_headers_only: tuple[Path, ...], new_headers_only: tuple[Path, ...],
-    old_includes_only: tuple[Path, ...], new_includes_only: tuple[Path, ...],
-) -> None:
-    """Validate appcompat CLI argument combinations."""
-    if weak_mode and (old_lib is not None or new_lib is not None):
-        raise click.UsageError(
-            "--check-against cannot be used with positional OLD_LIB/NEW_LIB arguments."
-        )
-    if not weak_mode and (old_lib is None or new_lib is None):
-        raise click.UsageError(
-            "Provide OLD_LIB and NEW_LIB arguments, or use --check-against for weak mode."
-        )
-    if weak_mode or list_symbols:
-        _rejected: list[str] = []
-        if old_headers_only:
-            _rejected.append("--old-header")
-        if new_headers_only:
-            _rejected.append("--new-header")
-        if old_includes_only:
-            _rejected.append("--old-include")
-        if new_includes_only:
-            _rejected.append("--new-include")
-        if _rejected:
-            mode_label = "--check-against" if weak_mode else "--list-required-symbols"
-            raise click.UsageError(
-                f"{', '.join(_rejected)} cannot be used with {mode_label}. "
-                f"Per-side header/include flags are only supported in full "
-                f"comparison mode (OLD_LIB NEW_LIB)."
-            )
-
-
-def _handle_list_required_symbols(
-    app_path: Path,
-    check_against_lib: Path | None,
-    old_lib: Path | None, new_lib: Path | None,
-    weak_mode: bool, fmt: str,
-    _get_lib_soname: Callable[[Path], str], parse_app_requirements: Callable[..., AppRequirements],
-) -> None:
-    """Handle the --list-required-symbols flow."""
-    target_lib = check_against_lib if weak_mode else (old_lib or new_lib)
-    if target_lib is None:
-        raise click.UsageError(
-            "--list-required-symbols requires a library path "
-            "(via positional args or --check-against)."
-        )
-    lib_name = _get_lib_soname(target_lib)
-    reqs = parse_app_requirements(app_path, lib_name)
-    if fmt == "json":
-        import json as _json
-        click.echo(_json.dumps({
-            "application": str(app_path),
-            "library": lib_name,
-            "needed_libs": reqs.needed_libs,
-            "required_symbols": sorted(reqs.undefined_symbols),
-            "required_versions": reqs.required_versions,
-        }, indent=2))
-    else:
-        click.echo(f"Application: {app_path}")
-        click.echo(f"Library filter: {lib_name}")
-        click.echo(f"Needed libraries: {', '.join(reqs.needed_libs) or '(none)'}")
-        click.echo(f"Required symbols ({len(reqs.undefined_symbols)}):")
-        for sym in sorted(reqs.undefined_symbols):
-            click.echo(f"  {sym}")
-        if reqs.required_versions:
-            click.echo(f"Required versions ({len(reqs.required_versions)}):")
-            for ver, lib in sorted(reqs.required_versions.items()):
-                click.echo(f"  {ver} (from {lib})")
-
-
-@main.command("appcompat")
-@click.argument("app_path", type=click.Path(exists=True, path_type=Path))
-@click.argument("old_lib", type=click.Path(exists=True, path_type=Path), required=False, default=None)
-@click.argument("new_lib", type=click.Path(exists=True, path_type=Path), required=False, default=None)
-# ── Weak mode ─────────────────────────────────────────────────────────────────
-@click.option("--check-against", "check_against_lib",
-              type=click.Path(exists=True, path_type=Path), default=None,
-              help="Weak mode: check if a library provides everything the app needs "
-                   "(no old library required).")
-# ── Dump options ──────────────────────────────────────────────────────────────
-@click.option("-H", "--header", "headers", multiple=True,
-              type=click.Path(path_type=Path),
-              help="Public header file or directory for library ABI extraction "
-                   "(applied to both sides).")
-@click.option("-I", "--include", "includes", multiple=True,
-              type=click.Path(path_type=Path),
-              help="Extra include directory for castxml (applied to both sides).")
-@click.option("--lang", default="c++", show_default=True,
-              type=click.Choice(["c++", "c"], case_sensitive=False),
-              help="Language mode for castxml.")
-@click.option("--old-header", "old_headers_only", multiple=True,
-              type=click.Path(path_type=Path),
-              help="Public header for old library only (overrides -H for old).")
-@click.option("--new-header", "new_headers_only", multiple=True,
-              type=click.Path(path_type=Path),
-              help="Public header for new library only (overrides -H for new).")
-@click.option("--old-include", "old_includes_only", multiple=True,
-              type=click.Path(path_type=Path),
-              help="Include dir for old library only (overrides -I for old).")
-@click.option("--new-include", "new_includes_only", multiple=True,
-              type=click.Path(path_type=Path),
-              help="Include dir for new library only (overrides -I for new).")
-@click.option("--old-version", "old_version", default="old", show_default=True)
-@click.option("--new-version", "new_version", default="new", show_default=True)
-# ── Output options ────────────────────────────────────────────────────────────
-@click.option("--format", "fmt", type=click.Choice(["json", "markdown", "html"]),
-              default="markdown", show_default=True)
-@click.option("-o", "--output", type=click.Path(path_type=Path), default=None)
-@click.option("--show-irrelevant", is_flag=True, default=False,
-              help="Include library changes that don't affect the application.")
-@click.option("--list-required-symbols", "list_symbols", is_flag=True, default=False,
-              help="List symbols the application requires and exit.")
-# ── Suppression + policy ─────────────────────────────────────────────────────
-@click.option("--suppress", type=click.Path(exists=True, path_type=Path), default=None,
-              help="Suppression file (YAML).")
-@click.option("--policy", "policy",
-              type=click.Choice(["strict_abi", "sdk_vendor", "plugin_abi"], case_sensitive=True),
-              default="strict_abi", show_default=True)
-@click.option("--policy-file", "policy_file_path",
-              type=click.Path(exists=True, path_type=Path), default=None)
-@click.option("-v", "--verbose", is_flag=True, default=False)
-def appcompat_cmd(
-    app_path: Path,
-    old_lib: Path | None,
-    new_lib: Path | None,
-    check_against_lib: Path | None,
-    headers: tuple[Path, ...],
-    includes: tuple[Path, ...],
-    lang: str,
-    old_headers_only: tuple[Path, ...],
-    new_headers_only: tuple[Path, ...],
-    old_includes_only: tuple[Path, ...],
-    new_includes_only: tuple[Path, ...],
-    old_version: str,
-    new_version: str,
-    fmt: str,
-    output: Path | None,
-    show_irrelevant: bool,
-    list_symbols: bool,
-    suppress: Path | None,
-    policy: str,
-    policy_file_path: Path | None,
-    verbose: bool,
-) -> None:
-    """Check if an application is compatible with a library update.
-
-    Answers: "Will my application still work with the new library version?"
-    by intersecting the app's required symbols with the library diff.
-
-    \b
-    Full check (with old and new library):
-      abicheck appcompat myapp libfoo.so.1 libfoo.so.2
-      abicheck appcompat myapp libfoo.so.1 libfoo.so.2 -H include/foo.h
-
-    \b
-    Weak mode (only new library — symbol availability check):
-      abicheck appcompat myapp --check-against libfoo.so.2
-
-    \b
-    List required symbols:
-      abicheck appcompat myapp --list-required-symbols --check-against libfoo.so.2
-
-    \b
-    Exit codes:
-      0  COMPATIBLE — application is safe with the new library
-      2  API_BREAK — source-level break affecting app's symbols
-      4  BREAKING — binary ABI break or missing symbols
-    """
-    _setup_verbosity(verbose)
-
-    from .appcompat import _get_lib_soname, check_appcompat, parse_app_requirements
-    from .appcompat import check_against as _check_against
-    from .reporter import appcompat_to_json, appcompat_to_markdown
-
-    weak_mode = check_against_lib is not None
-    _validate_appcompat_args(
-        weak_mode, old_lib, new_lib, list_symbols,
-        old_headers_only, new_headers_only,
-        old_includes_only, new_includes_only,
-    )
-
-    if list_symbols:
-        _handle_list_required_symbols(
-            app_path, check_against_lib, old_lib, new_lib,
-            weak_mode, fmt,
-            _get_lib_soname, parse_app_requirements,
-        )
-        return
-
-    if weak_mode:
-        assert check_against_lib is not None
-        result = _check_against(app_path, check_against_lib)
-    else:
-        assert old_lib is not None and new_lib is not None
-        suppression, pf = _load_suppression_and_policy(suppress, policy, policy_file_path)
-        old_h, new_h, old_inc, new_inc = _resolve_per_side_options(
-            headers, includes,
-            old_headers_only, new_headers_only,
-            old_includes_only, new_includes_only,
-        )
-        resolved_old_h = _expand_header_inputs(old_h) if old_h else []
-        resolved_new_h = _expand_header_inputs(new_h) if new_h else []
-        result = check_appcompat(
-            app_path, old_lib, new_lib,
-            old_headers=resolved_old_h,
-            new_headers=resolved_new_h,
-            old_includes=old_inc,
-            new_includes=new_inc,
-            old_version=old_version,
-            new_version=new_version,
-            lang=lang,
-            suppression=suppression,
-            policy=policy,
-            policy_file=pf,
-        )
-
-    if fmt == "json":
-        text = appcompat_to_json(result)
-    elif fmt == "html":
-        from .appcompat_html import appcompat_to_html
-        text = appcompat_to_html(result)
-    else:
-        text = appcompat_to_markdown(result, show_irrelevant=show_irrelevant)
-
-    if output:
-        output.write_text(text, encoding="utf-8")
-        click.echo(f"Report written to {output}", err=True)
-    else:
-        click.echo(text)
-
-    if result.verdict == Verdict.BREAKING:
-        sys.exit(4)
-    elif result.verdict == Verdict.API_BREAK:
-        sys.exit(2)
-
-@main.command("suggest-suppressions")
-@click.argument("diff_json", type=click.Path(exists=True, path_type=Path))
-@click.option("-o", "--output", type=click.Path(path_type=Path), default=None,
-              help="Output file for candidate suppressions (default: stdout).")
-@click.option("--expiry-days", type=click.IntRange(min=0), default=180, show_default=True,
-              help="Number of days from today for the expires field.")
-def suggest_suppressions_cmd(
-    diff_json: Path,
-    output: Path | None,
-    expiry_days: int,
-) -> None:
-    """Generate candidate suppression rules from a JSON diff result.
-
-    DIFF_JSON is a JSON file produced by 'abicheck compare --format json'.
-
-    \b
-    Example:
-      abicheck compare old.so new.so -H include/ --format json -o diff.json
-      abicheck suggest-suppressions diff.json -o candidates.yml
-    """
-
-    from .suppression import suggest_suppressions
-
-    try:
-        text = diff_json.read_text(encoding="utf-8")
-        data = json.loads(text)
-    except (OSError, json.JSONDecodeError) as e:
-        raise click.ClickException(f"Cannot read JSON diff: {e}") from e
-
-    if not isinstance(data, dict):
-        raise click.ClickException(
-            "JSON diff must be an object with a 'changes' key"
-        )
-    if "changes" not in data:
-        raise click.ClickException(
-            "JSON diff is missing required 'changes' key"
-        )
-    changes = data["changes"]
-    if not isinstance(changes, list):
-        raise click.ClickException("'changes' must be an array")
-    for i, entry in enumerate(changes):
-        if not isinstance(entry, dict):
-            raise click.ClickException(
-                f"changes[{i}] must be an object, got {type(entry).__name__}"
-            )
-
-    yaml_text = suggest_suppressions(changes, expiry_days=expiry_days)
-    _write_or_echo(output, yaml_text)
-
-
-# ── Full-stack dependency commands ────────────────────────────────────────────
-
-@main.command("deps")
-@click.argument("binary", type=click.Path(exists=True, path_type=Path))
-@click.option("--search-path", "search_paths", multiple=True,
-              type=click.Path(exists=True, path_type=Path),
-              help="Additional directory to search for shared libraries.")
-@click.option("--sysroot", type=click.Path(exists=True, path_type=Path), default=None,
-              help="Sysroot prefix for cross/container analysis.")
-@click.option("--ld-library-path", "ld_library_path", default="",
-              help="Simulated LD_LIBRARY_PATH (colon-separated).")
-@click.option("--format", "fmt", type=click.Choice(["json", "markdown", "html"]),
-              default="markdown", show_default=True)
-@click.option("-o", "--output", type=click.Path(path_type=Path), default=None)
-@click.option("-v", "--verbose", is_flag=True, default=False)
-def deps_cmd(
-    binary: Path, search_paths: tuple[Path, ...],
-    sysroot: Path | None, ld_library_path: str,
-    fmt: str, output: Path | None, verbose: bool,
-) -> None:
-    """Show the resolved dependency tree and symbol binding status.
-
-    Resolves the transitive closure of DT_NEEDED dependencies for BINARY
-    using loader-accurate search order (RPATH/RUNPATH, LD_LIBRARY_PATH,
-    default dirs) and reports symbol binding status.
-
-    \b
-    Exit codes:
-      0  All dependencies resolved, all required symbols bound
-      1  Missing dependencies or symbols (load would fail)
-
-    \b
-    Examples:
-      abicheck deps ./build/libfoo.so
-      abicheck deps /usr/bin/myapp --format json -o deps.json
-      abicheck deps ./app --sysroot /path/to/container/rootfs
-    """
-    _setup_verbosity(verbose)
-
-    fmt_detected = _detect_binary_format(binary)
-    if fmt_detected != "elf":
-        raise click.ClickException(
-            f"deps requires an ELF binary; got {fmt_detected or 'unknown format'}: {binary}"
-        )
-
-    from .stack_checker import check_single_env
-    from .stack_report import stack_to_json, stack_to_markdown
-
-    result = check_single_env(
-        binary,
-        search_paths=list(search_paths) or None,
-        sysroot=sysroot,
-        ld_library_path=ld_library_path,
-    )
-
-    if fmt == "json":
-        text = stack_to_json(result)
-    elif fmt == "html":
-        from .stack_html import stack_to_html
-        text = stack_to_html(result)
-    else:
-        text = stack_to_markdown(result)
-    if output:
-        _safe_write_output(output, text)
-        click.echo(f"Report written to {output}", err=True)
-    else:
-        click.echo(text)
-
-    if result.loadability.value == "fail":
-        sys.exit(1)
-
-
-@main.command("stack-check")
-@click.argument("binary", type=click.Path(path_type=Path))
-@click.option("--baseline", type=click.Path(exists=True, path_type=Path),
-              default=Path("/"), show_default=True,
-              help="Sysroot for the baseline environment.")
-@click.option("--candidate", type=click.Path(exists=True, path_type=Path),
-              default=Path("/"), show_default=True,
-              help="Sysroot for the candidate environment.")
-@click.option("--search-path", "search_paths", multiple=True,
-              type=click.Path(exists=True, path_type=Path),
-              help="Additional directory to search for shared libraries.")
-@click.option("--ld-library-path", "ld_library_path", default="",
-              help="Simulated LD_LIBRARY_PATH (colon-separated).")
-@click.option("--format", "fmt", type=click.Choice(["json", "markdown", "html"]),
-              default="markdown", show_default=True)
-@click.option("-o", "--output", type=click.Path(path_type=Path), default=None)
-@click.option("-v", "--verbose", is_flag=True, default=False)
-def stack_check_cmd(
-    binary: Path, baseline: Path, candidate: Path,
-    search_paths: tuple[Path, ...], ld_library_path: str,
-    fmt: str, output: Path | None, verbose: bool,
-) -> None:
-    """Compare a binary's full dependency stack across two environments.
-
-    Resolves all transitive dependencies in both BASELINE and CANDIDATE sysroots,
-    computes symbol bindings, detects changed DSOs, runs per-library ABI diffs,
-    and produces a stack-level compatibility verdict.
-
-    BINARY is the path relative to the sysroot (e.g. usr/bin/myapp).
-
-    \b
-    Exit codes:
-      0  PASS — binary loads and no harmful ABI changes
-      1  WARN — loads but ABI risk detected
-      4  FAIL — load failure or binary ABI break
-
-    \b
-    Examples:
-      abicheck stack-check usr/bin/myapp --baseline /old-root --candidate /new-root
-      abicheck stack-check usr/lib/libfoo.so.1 \\
-        --baseline ./image-v1 --candidate ./image-v2 --format json
-    """
-    _setup_verbosity(verbose)
-
-    # Guard against accidental no-op comparisons.
-    if baseline == candidate:
-        raise click.UsageError(
-            "--baseline and --candidate resolve to the same sysroot; "
-            "provide two different roots for stack comparison."
-        )
-
-    # Validate that every existing binary is ELF in both sysroots
-    for label, root in [("baseline", baseline), ("candidate", candidate)]:
-        resolved = root / binary
-        if resolved.exists():
-            fmt_detected = _detect_binary_format(resolved)
-            if fmt_detected != "elf":
-                raise click.ClickException(
-                    f"stack-check requires an ELF binary; got "
-                    f"{fmt_detected or 'unknown format'}: {resolved}"
-                )
-
-    from .stack_checker import check_stack
-    from .stack_report import stack_to_json, stack_to_markdown
-
-    result = check_stack(
-        binary,
-        baseline_root=baseline,
-        candidate_root=candidate,
-        ld_library_path=ld_library_path,
-        search_paths=list(search_paths) or None,
-    )
-
-    if fmt == "json":
-        text = stack_to_json(result)
-    elif fmt == "html":
-        from .stack_html import stack_to_html
-        text = stack_to_html(result)
-    else:
-        text = stack_to_markdown(result)
-    if output:
-        _safe_write_output(output, text)
-        click.echo(f"Report written to {output}", err=True)
-    else:
-        click.echo(text)
-
-    if result.loadability.value == "fail" or result.abi_risk.value == "fail":
-        sys.exit(4)
-    elif result.abi_risk.value == "warn" or result.loadability.value == "warn":
-        sys.exit(1)
-
 
 # ── ABICC compat subcommands (implementation in abicheck.compat) ─────────────
 # NOTE: eagerly loads abicheck.compat.cli at import time — intentional so all
@@ -1938,9 +1484,12 @@ main.add_command(compat_group)
 # sibling files to keep this module under the AI-readiness file-size limit.
 # ---------------------------------------------------------------------------
 from . import (  # noqa: E402  — must run after `main` and helpers are defined
+    cli_appcompat,  # noqa: F401  — registers appcompat
     cli_baseline,  # noqa: F401  — registers baseline
     cli_compare_release,  # noqa: F401  — registers compare-release
     cli_debian_symbols,  # noqa: F401  — registers debian-symbols
+    cli_stack,  # noqa: F401  — registers deps, stack-check
+    cli_suggest,  # noqa: F401  — registers suggest-suppressions
 )
 
 if __name__ == "__main__":
