+++ abicheck/diff_platform.py
@@ -20,9 +20,20 @@ from typing import Any
 from .checker_policy import ChangeKind
 from .checker_types import Change
 from .detector_registry import registry
+from .diff_platform_templates import (
+    _diff_template_inner_types as _diff_template_inner_types,
+)
+from .diff_platform_templates import (
+    _extract_template_args as _extract_template_args,
+)
+from .diff_platform_templates import (
+    _split_top_level_args as _split_top_level_args,
+)
+from .diff_platform_templates import (
+    _template_outer as _template_outer,
+)
 from .diff_symbols import _public_functions
 from .diff_types import _RESERVED_FIELD_RE
-from .dwarf_advanced import diff_advanced_dwarf
 from .elf_metadata import SymbolBinding, SymbolType
 from .model import (
     AbiSnapshot,
@@ -1362,39 +1373,6 @@ def _diff_enum_layouts(o: object, n: object) -> list[Change]:
     return changes
 
 
-# ── Sprint 4: Advanced DWARF (calling convention, toolchain flags, visibility) ─
-
-
-
-def _diff_advanced_dwarf(old: AbiSnapshot, new: AbiSnapshot) -> list[Change]:
-    """Sprint 4: calling convention, packing, toolchain flag drift."""
-    from .dwarf_advanced import AdvancedDwarfMetadata
-
-    o: AdvancedDwarfMetadata = getattr(old, "dwarf_advanced", None) or AdvancedDwarfMetadata()
-    n: AdvancedDwarfMetadata = getattr(new, "dwarf_advanced", None) or AdvancedDwarfMetadata()
-
-    _kind_map = {
-        "calling_convention_changed": ChangeKind.CALLING_CONVENTION_CHANGED,
-        "value_abi_trait_changed": ChangeKind.VALUE_ABI_TRAIT_CHANGED,
-        "struct_packing_changed": ChangeKind.STRUCT_PACKING_CHANGED,
-        "toolchain_flag_drift": ChangeKind.TOOLCHAIN_FLAG_DRIFT,
-        "type_visibility_changed": ChangeKind.TYPE_VISIBILITY_CHANGED,
-        "frame_register_changed": ChangeKind.FRAME_REGISTER_CHANGED,
-    }
-
-    return [
-        Change(
-            kind=_kind_map[kind_str],
-            symbol=sym,
-            description=desc,
-            old_value=old_val,
-            new_value=new_val,
-        )
-        for kind_str, sym, desc, old_val, new_val in diff_advanced_dwarf(o, n)
-        if kind_str in _kind_map
-    ]
-
-
 # ── PR #89: ELF fallback for = delete (issue #100) ───────────────────────────
 
 @registry.detector("elf_deleted_fallback")
@@ -1480,154 +1458,8 @@ def _diff_elf_deleted_fallback(old: AbiSnapshot, new: AbiSnapshot) -> list[Chang
 
 
 # ── PR #89: Template inner-type deep analysis (issues #38 / #73) ─────────────
-
-def _split_top_level_args(inner: str) -> list[str]:
-    """Split a template argument string on top-level commas.
-
-    Respects nested ``<>``, ``()``, ``[]``, and ``{}`` delimiters so that
-    types like ``std::function<void(int, double)>`` are not split incorrectly.
-    """
-    _OPEN = {"<": 0, "(": 1, "[": 2, "{": 3}  # pylint: disable=invalid-name
-    _CLOSE = {">": 0, ")": 1, "]": 2, "}": 3}  # pylint: disable=invalid-name
-
-    args: list[str] = []
-    current: list[str] = []
-    nesting = [0, 0, 0, 0]  # angle, paren, bracket, brace
-
-    for c in inner:
-        if c in _OPEN:
-            nesting[_OPEN[c]] += 1
-            current.append(c)
-        elif c == ">" and all(n == 0 for n in nesting[1:]) and nesting[0] > 0:
-            nesting[0] -= 1
-            current.append(c)
-        elif c in _CLOSE and c != ">":
-            nesting[_CLOSE[c]] -= 1
-            current.append(c)
-        elif c == "," and all(n == 0 for n in nesting):
-            args.append("".join(current).strip())
-            current = []
-        else:
-            current.append(c)
-    if current:
-        args.append("".join(current).strip())
-    return args
-
-
-def _extract_template_args(type_str: str) -> list[str] | None:
-    """Extract template argument string(s) from a type like ``vector<int>``.
-
-    Returns a list of top-level template arguments (splitting on ``,`` while
-    respecting nested ``<>``), or ``None`` if the type is not a template.
-
-    Examples::
-
-        "std::vector<int>"         → ["int"]
-        "std::map<int, double>"    → ["int", "double"]
-        "Foo<Bar<int>, double>"    → ["Bar<int>", "double"]
-        "int"                      → None
-        "std::vector<>"            → []
-    """
-    lt = type_str.find("<")
-    if lt == -1:
-        return None
-    # Find the matching closing >
-    depth = 0
-    for i, ch in enumerate(type_str[lt:], start=lt):
-        if ch == "<":
-            depth += 1
-        elif ch == ">":
-            depth -= 1
-            if depth == 0:
-                inner = type_str[lt + 1 : i].strip()
-                if not inner:
-                    return []
-                return _split_top_level_args(inner)
-    return None  # unbalanced brackets — skip
-
-
-def _template_outer(type_str: str) -> str:
-    """Return the outer template name, e.g. ``std::vector`` from ``std::vector<int>``."""
-    lt = type_str.find("<")
-    return type_str[:lt].rstrip() if lt != -1 else type_str
-
-
-@registry.detector("template_inner_types")
-def _diff_template_inner_types(old: AbiSnapshot, new: AbiSnapshot) -> list[Change]:
-    """Detect ABI-relevant template inner-type changes in function signatures.
-
-    Compares param types and return types for functions present in both snapshots.
-    When both old and new have a template specialization (e.g. ``std::vector<T>``)
-    with the *same outer template name* but *different type arguments*, this is an
-    ABI break: the instantiation's layout, size, and ABI fingerprint all differ.
-
-    This detector fires in addition to FUNC_PARAMS_CHANGED / FUNC_RETURN_CHANGED
-    to provide a more specific, actionable description of the inner-type change.
-
-    Example::
-
-        void process(std::vector<int> v)   →   void process(std::vector<double> v)
-        # → TEMPLATE_PARAM_TYPE_CHANGED: "std::vector" inner type int → double
-
-    NOTE on mangling: Under the Itanium C++ ABI, parameter types ARE included in the
-    mangled symbol name, so a real ``std::vector<int>`` → ``std::vector<double>`` param
-    change produces different mangled names (FUNC_REMOVED + FUNC_ADDED, not an intersection
-    hit). This detector therefore only fires for:
-      1. Return type template changes (return type is NOT in Itanium mangling for
-         non-template functions, so the mangled name stays the same).
-      2. Cases where the snapshot was produced with simplified/un-mangled names (e.g.
-         from header-only analysis without a compiled .so).
-    For production ELF-based snapshots, FUNC_PARAMS_CHANGED is the primary signal.
-    """
-    changes: list[Change] = []
-    old_map = _public_functions(old)
-    new_map = _public_functions(new)
-
-    for mangled in set(old_map) & set(new_map):
-        f_old = old_map[mangled]
-        f_new = new_map[mangled]
-
-        # --- Return type template inner change ---
-        old_ret_args = _extract_template_args(f_old.return_type)
-        new_ret_args = _extract_template_args(f_new.return_type)
-        if (
-            old_ret_args is not None
-            and new_ret_args is not None
-            and old_ret_args != new_ret_args
-            and _template_outer(f_old.return_type) == _template_outer(f_new.return_type)
-        ):
-            changes.append(Change(
-                kind=ChangeKind.TEMPLATE_RETURN_TYPE_CHANGED,
-                symbol=mangled,
-                description=(
-                    f"Template return type inner argument changed: {f_old.name} "
-                    f"({f_old.return_type} → {f_new.return_type})"
-                ),
-                old_value=f_old.return_type,
-                new_value=f_new.return_type,
-            ))
-
-        # --- Param template inner change ---
-        for i, (p_old, p_new) in enumerate(zip(f_old.params, f_new.params)):
-            old_args = _extract_template_args(p_old.type)
-            new_args = _extract_template_args(p_new.type)
-            if (
-                old_args is not None
-                and new_args is not None
-                and old_args != new_args
-                and _template_outer(p_old.type) == _template_outer(p_new.type)
-            ):
-                param_label = p_old.name or str(i)
-                changes.append(Change(
-                    kind=ChangeKind.TEMPLATE_PARAM_TYPE_CHANGED,
-                    symbol=mangled,
-                    description=(
-                        f"Template parameter inner type changed: {f_old.name} "
-                        f"param {param_label} ({p_old.type} → {p_new.type})"
-                    ),
-                    old_value=p_old.type,
-                    new_value=p_new.type,
-                ))
-
-    return changes
+# Detectors moved to ``diff_platform_templates`` to keep this file under the
+# AI-readiness file-size soft cap. Re-exported below so existing imports from
+# ``abicheck.diff_platform`` keep working.
+# Re-exports for backwards compatibility — see top-of-file imports.
